pyproject.toml
[project]
name = "pdf-agent-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "langchain>=1.0.2",
    "langchain-community>=0.4",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.1",
    "pdf2image>=1.17.0",
    "pi-heif>=1.1.1",
    "pypdf>=6.1.3",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "tesseract>=0.1.3",
]
